requirements.txt
selenium
pytest
webdriver_manager
py
nose
pytest-xdist
python-dotenv